pyproject.toml
[tool.poetry]
name = "repl_python3_Classroom-Discord-Bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.7.3"
Flask = "^2.0.1"
pytz = "^2021.1"
PyNaCl = "^1.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
